requirements.txt
selenium == 4.38.0
robotframework == 7.3.2
robotframework-seleniumlibrary == 6.8.0
behave==1.2.6
webdriver-manager==4.0.1
pytest==6.2.4
